pyproject.toml
[project]
name = "my-cool-project"
version = "0.1.0"
dependencies = [
    "astroid==3.3.9",
    "bleach>=6.3.0",
    "blinker==1.9.0",
    "certifi==2025.4.26",
    "charset-normalizer==3.4.1",
    "click==8.1.8",
    "colorama==0.4.6",
    "contourpy==1.3.2",
    "cycler==0.12.1",
    "dill==0.4.0",
    "et-xmlfile==2.0.0",
    "filelock==3.18.0",
    "flask==3.1.0",
    "flask-cors==6.0.1",
    "fonttools==4.57.0",
    "fsspec==2025.5.1",
    "gunicorn==23.0.0",
    "idna==3.10",
    "iniconfig==2.1.0",
    "isort==6.0.1",
    "itsdangerous==2.2.0",
    "jinja2==3.1.6",
    "kiwisolver==1.4.8",
    "llvmlite==0.44.0",
    "lxml==5.4.0",
    "markupsafe==3.0.2",
    "matplotlib==3.10.1",
    "mccabe==0.7.0",
    "more-itertools==10.7.0",
    "mpmath==1.3.0",
    "networkx==3.5",
    "numba==0.61.2",
    "numpy==2.2.5",
    "openai-whisper==20240930",
    "openpyxl==3.1.5",
    "packaging==25.0",
    "pandas==2.2.3",
    "pillow==11.2.1",
    "platformdirs==4.3.7",
    "pluggy==1.5.0",
    "py==1.11.0",
    "pycodestyle==2.13.0",
    "pyflakes==3.3.2",
    "pylint==3.3.6",
    "pyparsing==3.2.3",
    "pytest==8.3.5",
    "pytest-pycodestyle==2.4.1",
    "python-dateutil==2.9.0.post0",
    "python-docx==1.1.2",
    "python-dotenv==1.1.0",
    "pytz==2025.2",
    "regex==2024.11.6",
    "requests==2.32.3",
    "scipy==1.15.3",
    "setuptools==80.9.0",
    "six==1.17.0",
    "sympy==1.14.0",
    "testfixtures==8.3.0",
    "tiktoken==0.9.0",
    "tomlkit==0.13.2",
    "torch==2.7.1",
    "tqdm==4.67.1",
    "typing-extensions==4.13.2",
    "tzdata==2025.2",
    "urllib3==2.4.0",
    "werkzeug==3.1.3",
    "xlsxwriter==3.2.3",
]
